requirements.txt
django>=5.2.1
djangorestframework>=3.16.1
pillow>=12.0.0
python-dotenv
python-decouple>=3.8
celery[redis]>=5.3.0
redis>=5.0.0
django-redis>=5.4.0
hiredis>=2.3.2
django-celery-beat
django-celery-results
djangorestframework-simplejwt
whitenoise>=6.6.0
gunicorn>=21.2.0
uvicorn[standard]>=0.29.0
uvloop>=0.19.0
psycopg2-binary>=2.9.9
